+++ faccessat.c
@@ -45,8 +45,8 @@ int faccessat(int fd, const char *path, int mode, int flags)
 		return -1;
 	}
 
-	__debug("%s(fd: %i, path: '%s' -> '%s', flags: 0x%x)\n", __func__, fd,
-		path, buf, flags);
+	__debug("%s(fd: %i, path: '%s' -> '%s', mode: 0%03o, flags: 0x%x)\n",
+		__func__, fd, path, buf, mode, flags);
 
 	__remove_at_empty_path_if_needed(buf, flags);
 	return next_faccessat(fd, buf, mode, flags);